pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "html2docx-converter"
version = "0.1.0"
description = "HTML to Word document converter with CSS variable resolution, bookmarks, and formatting inheritance."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Protectyr Labs", email = "labs@protectyr.com"},
]
keywords = ["html", "docx", "word", "converter", "css"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Markup :: HTML",
]

dependencies = [
    "beautifulsoup4>=4.12",
    "python-docx>=1.0",
    "lxml>=4.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://github.com/protectyr-labs/html2docx"
Repository = "https://github.com/protectyr-labs/html2docx"
Issues = "https://github.com/protectyr-labs/html2docx/issues"

[tool.setuptools.packages.find]
include = ["html2docx*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
